refactor(stage4): make theme/renderer options a single source
Candidate D. AVAILABLE_THEMES, AVAILABLE_MATH_RENDERERS, their defaults, and
the validate_* functions were duplicated across render-copy-ready.py and
wechat_draft.py with drifting exception types (ValueError vs RuntimeError).

- woeai/wechat/options.py: the one definition of all four constants plus
  validate_theme/validate_math_renderer (unified to ValueError).
- render-copy-ready.py: imports from options; self-bootstraps sys.path.
- wechat_draft.py: imports from options (was the RuntimeError copy).

Note: the two math renderers are already two real adapters at a seam. A full
renderer registry interface is intentionally deferred - the string-compare
selection works, and the immediate win is eliminating the duplicated
allowed-values list. PNG tier (mentioned in CONTEXT.md) remains a future
adapter to register.

55 tests green.
